Make the background of non-normal client's parents ParentRelative
non-normal clients (desktop features, panels) might want to be be
pseudo-transparent by settign their background pixmap to ParentRelative.
This sets the packground pixmap of the ObFrame of such clients to
ParentRelative (instead of BlackPixel) so that this pseudo transparency
works.
We don't do this for all clients for performance (?) reasons.