Fix deprecation warnings
Most of this is replacing calls to g_memdup with ones to g_memdup2.
Also removed a useless call to rsvg_handle_close. This function only
needs to be used if writing to a handle manually with rsvg_handle_write,
not as we are doing from a file with rsvg_handle_new_from_file, which
blocks until the read is comlete and closes the handle itself.